Deprecate scrapy.utils.benchserver.
